refactor(api): move badge-criteria-repository
