+++ api/scripts/create-badge-criteria-for-specified-badge.js
@@ -7,7 +7,7 @@ import { disconnect } from '../db/knex-database-connection.js';
 import { NotFoundError } from '../lib/domain/errors.js';
 import { SCOPES } from '../lib/domain/models/BadgeDetails.js';
 import { DomainTransaction } from '../lib/infrastructure/DomainTransaction.js';
-import * as badgeCriteriaRepository from '../lib/infrastructure/repositories/badge-criteria-repository.js';
+import * as badgeCriteriaRepository from '../src/evaluation/infrastructure/repositories/badge-criteria-repository.js';
 import * as badgeRepository from '../src/shared/infrastructure/repositories/badge-repository.js';
 
 // Usage: node scripts/create-badge-criteria-for-specified-badge path/data.json
