+++ api/src/shared/domain/usecases/create-badge.js
@@ -1,5 +1,5 @@
-import { MissingBadgeCriterionError } from '../../../src/shared/domain/errors.js';
-import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
+import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { MissingBadgeCriterionError } from '../errors.js';
 
 const createBadge = async function ({
   targetProfileId,
