+++ api/src/shared/domain/usecases/index.js
@@ -1,6 +1,8 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
+import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
+import * as badgeCriteriaRepository from '../../../evaluation/infrastructure/repositories/badge-criteria-repository.js';
 import * as activityAnswerRepository from '../../../school/infrastructure/repositories/activity-answer-repository.js';
 import * as activityRepository from '../../../school/infrastructure/repositories/activity-repository.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
@@ -19,8 +21,10 @@ const dependencies = {
   activityAnswerRepository,
   activityRepository,
   assessmentRepository,
-  challengeRepository,
   badgeRepository,
+  badgeCriteriaRepository,
+  challengeRepository,
+  targetProfileRepository,
 };
 
 const sharedUsecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
